+++ main.py
@@ -2,13 +2,12 @@ print("Hello The Brave New World")
 
 import sys
 from Adafruit_IO import MQTTClient
-# from simple_ai import *
-from physical import *
-import time
+# from AI.simple_ai import *
+from IoT.physical import *
 
 AIO_FEED_ID = ["sensor_1", "sensor_2", "sensor_3", "actuator_1", "actuator_2", "vision_detection"]
 AIO_USERNAME = "truonghuy"
-AIO_KEY = "aio_PXVN21tAWJcKoCCthvhVAnu7mish"
+AIO_KEY = "aio_UCMX70XwlAPnumsp4BVKNjcu8jvL"
 
 def connected(client):
     print("Ket noi thanh cong ...")
@@ -32,24 +31,17 @@ def message(client , feed_id , payload):
         setDevice2(state)
 
 client = MQTTClient(AIO_USERNAME , AIO_KEY)
-client.on_connect = connected
+client.on_connect = connected # passing function as object
 client.on_disconnect = disconnected
 client.on_message = message
 client.on_subscribe = subscribe
-client.connect()
-client.loop_background()
-#
-# def set_state(client, feed_id, payload):
-#     state = (str(payload) == "1") ? True: False
-#     if feed_id == "actuator_1":
-#         setDevice1(state)
-#     if feed_id == "actuator_2":
-#         setDevice2(state)
+client.connect() # call connect()
+client.loop_background() # call loop_background()
 
 while True:
     pass
 
-
+# #Simple AI Code
 # while True:
 #     time.sleep(5)
 #     image_capture()
